requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asn1crypto==1.5.1
asyncio==4.0.0
attrs==25.4.0
betterproto==2.0.0b7
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dotenv==0.9.9
frozenlist==1.8.0
grpclib==0.4.9
h2==4.3.0
hpack==4.1.0
http_ece==1.2.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
numpy==2.4.1
oscrypto==1.3.0
pillow==12.1.0
propcache==0.4.1
protobuf==6.33.4
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rustPlusPushReceiver==0.6.1
scipy==1.17.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yarl==1.22.0
